requirements.txt
# Web framework
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary

# ML/Matching
numpy
pandas
scikit-learn

# SMS (Africa's Talking)
requests
certifi

# Utilities
python-dotenv
click
fastapi
asyncpg

